requirements.txt
Flask==3.1.0
gunicorn==20.1.0
Werkzeug==3.1.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
